Wed, Oct 5, 2022, 6:15 PM +00:00

+++ filter-more/helpers.c
@@ -54,9 +54,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
         //Next Pixel
         for (int j = 0; j < width; j++)
         {
-            int sumred = 0;
-            int sumgreen = 0;
-            int sumblue = 0;
+            float sumred = 0;
+            float sumgreen = 0;
+            float sumblue = 0;
             int pxlcount = 0;
             //Check adjacent pixels
             for (int k = -1; k < 2; k++)
